pdfbox/font: act on the slice 4 pull request feedback
Six comments: three port defects, one missing piece of scope, two Java
behaviours the port already reproduced.

CMap.toInt accumulated at 64 bits where Java's int is 32, so a four-byte code
with the top bit set stayed positive. The width is observable: toUnicode(int)
tests the code against 256, 0xFFFF and 0xFFFFFF to decide how many bytes it
had, so a negative code takes the two-byte branch and misses the mapping
stored under that same key. cmap/feedback_test.go pins the arithmetic and the
lookup it decides.

PDCIDFont.readVerticalDisplacements had softened Java's (COSNumber) casts to a
warning and a partial read, so a malformed /W2 gave a font with half its
vertical metrics. The casts are back; the panic lands where Java throws.

KerningTable.read now narrows its version 1 count the way Java's cast does.
The branch is unreachable — case 1 cannot match a version that is 0 or at
least 0x10000 — which is JAVA-BUGS 22, so no Apple-format kern table is ever
read.

PDFontFactory.createFont was missing the Type 0 /Subtype repair, deferred
while PDType0Font was unported and never lifted; the stale header comment is
what the review noticed. fixType0Subtype, getFontTypeFromFont, getFontHeader,
the six header sniffers and the FontType inner class are ported, along with
the two log lines the port had dropped.

Declined, both recorded: CMapStrings.getMapping reading a zero-length code as
the two-byte code 0 is what Java does (JAVA-BUGS 23), and
createDescendantFont naming the /Type rather than the /Subtype in its error is
Java's message, now reproduced down to the COSName{...} it concatenates.

Corpus unchanged at 34 of 40 unsorted, 33 sorted.

+++ go/fontbox/ttf/kerning.go
@@ -31,7 +31,14 @@ func (t *KerningTable) Read(ttf *TrueTypeFont, data DataStream) error {
 	case 0:
 		numSubtables = r.unsignedShort()
 	case 1:
-		numSubtables = int(r.unsignedInt())
+		// Java narrows the unsigned count to a signed 32-bit int, so a count
+		// with bit 31 set goes negative and the check below skips it; a Go int
+		// is 64 bits and would keep it positive and allocate on it.
+		//
+		// JAVA-BUGS entry 22: this case cannot be reached. version is zero or
+		// at least 0x10000 by the time the switch sees it, so an Apple
+		// version 1.0 'kern' table falls through to the default and is skipped.
+		numSubtables = int(int32(r.unsignedInt()))
 	default:
 		slog.Debug("Skipped kerning table due to an unsupported kerning table version",
 			"version", version)
